abastecimento atualizado

+++ src/main/java/com/br/cooapi/abasteci/AbasteciController.java
@@ -19,7 +19,7 @@ public class AbasteciController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<AbasteciDto> findById (@PathVariable int idAbasteci) {
+    public ResponseEntity<AbasteciDto> findById (@PathVariable Long idAbasteci) {
         AbasteciDto abasteci = service.findById(idAbasteci);
         return ResponseEntity.ok().body(abasteci);
     }
@@ -30,12 +30,12 @@ public class AbasteciController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<AbasteciDto> update (@PathVariable Integer id, @RequestBody AbasteciForm obj) {
+    public ResponseEntity<AbasteciDto> update (@PathVariable Long id, @RequestBody AbasteciForm obj) {
         return ResponseEntity.ok(service.update(id, obj));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete (@PathVariable int id) {
+    public ResponseEntity<Void> delete (@PathVariable Long id) {
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
